// reference-code/esr/esale-mobile/src/shared/components/suggestions/interface/task-suggest-interface.ts
export interface ITaskSuggestProps {
  invisibleLabel?: boolean, // hiden label
  typeSearch: number, // type search (SINGLE or MULTI)
  fieldLabel: string, // label of field
  isRelation?: boolean,//status relation
  isError?: boolean,// set background error
  isRequire?: boolean,// display require label
  initData?: TaskSuggest[],// initial data
  hiddenSelectedData?: boolean,// hidden data selected
  updateStateElement: (searchValue: any) => void; // callback when change status control
  exportError?: (err: any) => void;
}
// ...
export interface TaskSuggest {
  taskId: number,// data in database mapping data form
  parentCustomerName: string, // data in database mapping
  milestoneName: string,// data in database mapping
  customerName: string,// data in database mapping
  productTradingName: string,
  taskName: string,// data in database mapping
  finishDate: string,// data in database mapping
  finishType: number,// data in database mapping
  employeeName: string,// data in database mapping
  status: number,// data in database mapping
  data: any
}